refactor(utils): reorganize functions; move authors file resolution to utils
