+++ package.json
@@ -109,7 +109,7 @@
     "netlify-cli": "^8.0.5",
     "nodemon": "^2.0.13",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.5.0",
+    "prettier": "^2.5.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-test-renderer": "^17.0.2",
