+++ packages/create-docusaurus/templates/facebook/package.json
@@ -35,7 +35,7 @@
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-react": "^7.27.0",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "prettier": "^2.5.0",
+    "prettier": "^2.5.1",
     "stylelint": "^13.2.1"
   },
   "browserslist": {
