+++ packages/docusaurus-plugin-content-blog/src/index.ts
@@ -16,6 +16,8 @@ import {
   posixPath,
   addTrailingPathSeparator,
   createAbsoluteFilePathMatcher,
+  getContentPathList,
+  getDataFilePath,
   DEFAULT_PLUGIN_ID,
 } from '@docusaurus/utils';
 import {translateContent, getTranslationFiles} from './translations';
@@ -45,13 +47,11 @@ import {
 import {Configuration} from 'webpack';
 import {
   generateBlogPosts,
-  getContentPathList,
   getSourceToPermalink,
   getBlogTags,
 } from './blogUtils';
 import {BlogPostFrontMatter} from './blogFrontMatter';
 import {createBlogFeedFiles} from './feed';
-import {getAuthorsMapFilePath} from './authors';
 
 export default async function pluginContentBlog(
   context: LoadContext,
@@ -90,8 +90,8 @@ export default async function pluginContentBlog(
   const aliasedSource = (source: string) =>
     `~blog/${posixPath(path.relative(pluginDataDirRoot, source))}`;
 
-  const authorsMapFilePath = await getAuthorsMapFilePath({
-    authorsMapPath: options.authorsMapPath,
+  const authorsMapFilePath = await getDataFilePath({
+    filePath: options.authorsMapPath,
     contentPaths,
   });
