+++ packages/docusaurus-utils/src/__tests__/__fixtures__/dataFiles/actualData/bad.json
@@ -1,0 +1,1 @@
+{"a": 2}
