+++ packages/docusaurus-utils/src/__tests__/__fixtures__/dataFiles/actualData/valid.json
@@ -1,0 +1,1 @@
+{"a": 1}
