+++ packages/docusaurus-utils/src/__tests__/__fixtures__/dataFiles/actualData/valid.yml
@@ -1,0 +1,1 @@
+a: 1
