+++ packages/docusaurus-utils/src/index.ts
@@ -20,30 +20,21 @@ import {
 
 import resolvePathnameUnsafe from 'resolve-pathname';
 
-import {posixPath as posixPathImport} from './posixPath';
 import {simpleHash, docuHash} from './hashUtils';
-import {normalizeUrl} from './normalizeUrl';
 import {DEFAULT_PLUGIN_ID} from './constants';
 
 export * from './constants';
 export * from './mdxUtils';
-export * from './normalizeUrl';
+export * from './urlUtils';
 export * from './tags';
-
-export const posixPath = posixPathImport;
-
 export * from './markdownParser';
 export * from './markdownLinks';
-export * from './escapePath';
 export * from './slugger';
-export {md5Hash, simpleHash, docuHash} from './hashUtils';
-export {
-  Globby,
-  GlobExcludeDefault,
-  createMatcher,
-  createAbsoluteFilePathMatcher,
-} from './globUtils';
+export * from './pathUtils';
+export * from './hashUtils';
+export * from './globUtils';
 export * from './webpackUtils';
+export * from './dataFileUtils';
 
 const fileHash = new Map();
 export async function generate(
@@ -133,17 +124,6 @@ export function genComponentName(pagePath: string): string {
   return upperFirst(camelCase(pageHash));
 }
 
-// When you want to display a path in a message/warning/error,
-// it's more convenient to:
-// - make it relative to cwd()
-// - convert to posix (ie not using windows \ path separator)
-// This way, Jest tests can run more reliably on any computer/CI
-// on both Unix/Windows
-// For Windows users this is not perfect (as they see / instead of \) but it's probably good enough
-export function toMessageRelativeFilePath(filePath: string): string {
-  return posixPath(path.relative(process.cwd(), filePath));
-}
-
 const chunkNameCache = new Map();
 /**
  * Generate unique chunk name given a module path.
@@ -172,18 +152,6 @@ export function genChunkName(
   return chunkName;
 }
 
-// Too dynamic
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any
-export function idx(target: any, keyPaths?: string | (string | number)[]): any {
-  return (
-    target &&
-    keyPaths &&
-    (Array.isArray(keyPaths)
-      ? keyPaths.reduce((obj, key) => obj && obj[key], target)
-      : target[keyPaths])
-  );
-}
-
 /**
  * Given a filepath and dirpath, get the first directory.
  */
@@ -195,27 +163,6 @@ export function getSubFolder(file: string, refDir: string): string | null {
   );
   const match = regexSubFolder.exec(file);
   return match && match[1];
-}
-
-/**
- * Alias filepath relative to site directory, very useful so that we
- * don't expose user's site structure.
- * Example: some/path/to/website/docs/foo.md -> @site/docs/foo.md
- */
-export function aliasedSitePath(filePath: string, siteDir: string): string {
-  const relativePath = posixPath(path.relative(siteDir, filePath));
-  // Cannot use path.join() as it resolves '../' and removes
-  // the '@site'. Let webpack loader resolve it.
-  return `@site/${relativePath}`;
-}
-
-export function getEditUrl(
-  fileRelativePath: string,
-  editUrl?: string,
-): string | undefined {
-  return editUrl
-    ? normalizeUrl([editUrl, posixPath(fileRelativePath)])
-    : undefined;
 }
 
 export function isValidPathname(str: string): boolean {
@@ -306,7 +253,7 @@ export function getPluginI18nPath({
   );
 }
 
-export async function mapAsyncSequencial<T, R>(
+export async function mapAsyncSequential<T, R>(
   array: T[],
   action: (t: T) => Promise<R>,
 ): Promise<R[]> {
@@ -330,35 +277,6 @@ export async function findAsyncSequential<T>(
     }
   }
   return undefined;
-}
-
-// return the  first folder path in which the file exists in
-export async function findFolderContainingFile(
-  folderPaths: string[],
-  relativeFilePath: string,
-): Promise<string | undefined> {
-  return findAsyncSequential(folderPaths, (folderPath) =>
-    fs.pathExists(path.join(folderPath, relativeFilePath)),
-  );
-}
-
-export async function getFolderContainingFile(
-  folderPaths: string[],
-  relativeFilePath: string,
-): Promise<string> {
-  const maybeFolderPath = await findFolderContainingFile(
-    folderPaths,
-    relativeFilePath,
-  );
-  // should never happen, as the source was read from the FS anyway...
-  if (!maybeFolderPath) {
-    throw new Error(
-      `File "${relativeFilePath}" does not exist in any of these folders:\n- ${folderPaths.join(
-        '\n- ',
-      )}]`,
-    );
-  }
-  return maybeFolderPath;
 }
 
 export function reportMessage(
