+++ website/docusaurus.config.js
@@ -176,7 +176,7 @@ const config = {
         max: 1030, // max resized image's size.
         min: 640, // min resized image's size. if original is lower, use that size.
         steps: 2, // the max number of images generated between min and max (inclusive)
-        disableInDev: false,
+        // disableInDev: false,
       },
     ],
     [
